+++ agent/api-v1/ping_test.go
@@ -12,5 +12,4 @@ func TestPing(t *testing.T) {
 	resp, err := http.Get("http://127.0.0.1:7651/v1/status/ping")
 	assert.NoError(t, err)
 	assert.Equal(t, resp.StatusCode, 200)
-	t.Log("OK")
 }